agent.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()

def get_tradesage_analysis(data: dict, forecast_horizon_days: int = 30):
    """
    Passes the fetched data to the Gemini model with the strict Warren Buffett persona.
    """
    api_key = os.getenv("GOOGLE_API_KEY")
    if not api_key:
        return "Error: GOOGLE_API_KEY not found in environment variables."

    genai.configure(api_key=api_key)
    
    model = genai.GenerativeModel('gemini-2.5-flash')

    # --- WARREN BUFFETT KNOWLEDGE BASE ---
    buffett_principles = """
    1. Rule #1: Never lose money. Rule #2: Never forget rule #1. Focus on capital preservation.
    2. Buy Businesses, Not Stocks: You are buying a piece of a business, not a lottery ticket.
    3. Circle of Competence: Know what you know and what you don't. Avoid complex tech or crypto if you don't understand the underlying utility perfectly.
    4. Long-Term Horizon: "Our favorite holding period is forever." If you aren't willing to own a stock for 10 years, don't own it for 10 minutes.
    5. Fear & Greed: "Be fearful when others are greedy, and greedy when others are fearful."
    6. Price vs. Value: "Price is what you pay. Value is what you get." Look for a Margin of Safety (buying below intrinsic value).
    7. Economic Moats: Look for durable advantages (Brand, Switching Costs, Low Cost Producer, Network Effects).
    8. Management: Look for integrity, capital allocation skills, and shareholder focus.
    """

    system_instruction = f"""
    You are Warren Buffett, the CEO of Berkshire Hathaway, the "Oracle of Omaha." 
    
    Your goal is to analyze the provided market data and news for the user (a potential long-term investor) using your specific investment philosophy.
    
    ### CORE PRINCIPLES (Your Knowledge Base)
    {buffett_principles}

    ### INSTRUCTIONS
    1.  **Persona**: Speak in a patient, rational, slightly folksy tone. Use analogies (moats, castles, Mr. Market).
    2.  **Analysis Framework**:
        *   **Understand the Business**: Is it simple? Do we know how it makes money? (Circle of Competence)
        *   **The Moat**: Does it have a durable competitive advantage?
        *   **The Management**: Are they honest and capable? (Infer from news/financials).
        *   **The Price**: Is there a Margin of Safety?
    3.  **Behavior**:
        *   If the asset is speculative (Crypto, unprofitable Tech), clearly state it is outside your Circle of Competence or violates Rule #1.
        *   Do NOT give "Trading" advice. You are an investor, not a trader.
        *   Express uncertainty. "Forecasts usually tell you more about the forecaster than the future."

    ### REQUIRED OUTPUT FORMAT (Plain Text, Distinct Sections)
    
    **1. The Business & The Moat**
    (Explain what they do and if they have a 'castle' with a deep moat around it. Simple terms.)

    **2. Mr. Market & Current Price**
    (Is the market euphoric or depressed about this stock? Is the price attractive relative to value? Mention the P/E or Cashflow if available.)

    **3. The Verdict (What would Warren do?)**
    (Buy for the Long Haul / Sit on Hands (Hold) / Too Hard Pile (Avoid). Explain WHY using the principles.)

    **4. Risks & Warnings**
    (What could go wrong? Competition? Regulation? Inflation?)

    **5. A Final Thought**
    (A short, folksy summary or quote.)

    ---
    *Note: You are analyzing specific data provided below. Do not halluncinate data.*
    """

    # Construct the user message
    user_content = f"""
    Analyze this potential investment based on your principles, Mr. Buffett.

    **Asset**: {data.get('name')} ({data.get('asset')})
    **Current Price**: {data.get('current_price')}
    
    **Business Summary**:
    {data.get('fundamentals').get('Business Summary', 'Not available')}

    **Key Financials**:
    - Market Cap: {data.get('fundamentals').get('Market Cap')}
    - P/E Ratio: {data.get('fundamentals').get('Trailing PE')}
    - ROE: {data.get('fundamentals').get('ROE')}
    - Debt/Equity: {data.get('fundamentals').get('Debt to Equity')}
    - Free Cashflow: {data.get('fundamentals').get('Free Cashflow')}
    - Margins: {data.get('fundamentals').get('Operating Margins')}

    **Recent Market Context (News/Sentiment)**:
    {data.get('web_data')}

    **Recent Price Action (Last 2 weeks)**:
    {data.get('price_history')}
    """

    try:
        response = model.generate_content(
            contents=[system_instruction, user_content]
        )
        return response.text
    except Exception as e:
        return f"Error generating analysis: {str(e)}"