requirements.txt
streamlit
google-generativeai
yfinance
python-dotenv
pandas
exa-py
